% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximate_mode.R
\name{approximate_mode}
\alias{approximate_mode}
\title{approximate_mode}
\usage{
approximate_mode(x)
}
\arguments{
\item{x}{alpha values for plotting point estimate}
}
\value{
mode of alpha distribution
}
\description{
Description of what the function does.
}
\examples{
\dontrun{
approximate_mode(x)
}


}
