% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesForest.R
\name{BayesForest}
\alias{BayesForest}
\title{BayesForest}
\usage{
BayesForest(
  model,
  study_info,
  data,
  dens_color = "skyblue",
  point_color = "black",
  CI_offset = 0.75,
  overlap = 1,
  trim_percentile = 0.01,
  y_text_size = 12,
  y_title_size = 12,
  CI_size = 12,
  x_text_size = 12,
  x_title_size = 12
)
}
\arguments{
\item{model}{RJAGS Model}

\item{study_info}{Column with study labels}

\item{data}{Data frame with origianl data}

\item{dens_color}{Color of density curves}

\item{point_color}{Color of point estimates}

\item{CI_offset}{Distance of critical intervals from density curves}

\item{overlap}{Overlap between study density curves}

\item{trim_percentile}{The limits of the density curves}

\item{y_text_size}{Size of study labels}

\item{y_title_size}{Size of y-axis label}

\item{CI_size}{Size of critical interval text}

\item{x_text_size}{Size of x-axis tick text}

\item{x_title_size}{Size of x-axis title text}
}
\value{
Forest plots of density curves
}
\description{
Creates forest plot of effect size estimates at the study level with distributions
}
\examples{

BayesForest(model, study_info = "Study_Info",
data = data, dens_color = "skyblue",
point_color = "black", overlap = 1,
x_title_size = 14, y_title_size = 14)






}
