% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAGS.summary.R
\name{JAGS.summary}
\alias{JAGS.summary}
\title{JAGS.summary}
\usage{
JAGS.summary(mcmc_samples)
}
\arguments{
\item{mcmc_samples}{A RJAGS model Object}
}
\value{
A simplified model summary that excludes estimates at the study level
}
\description{
Creates a simplified model summary output
}
\examples{

mod_1 <- BayesMeta(formula = yi ~ 1+predictor1, v = "vi", studyID = "study_id",
  data = outlier_rm_test, n_chains = 5, n_updates = 10000, n_iterations = 10000,
  prob_conditions = list(
    "mu.alpha" = list("greater" = c(0.2, 0.5),"lesser" = c(-0.2, -0.5), "between" = list(c(-.2,.3))),
    "sd.alpha" = list("greater" = c(0.2, 0.3))))
}
