% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAGSmeta.R
\name{JAGSmeta}
\alias{JAGSmeta}
\title{JAGSmeta}
\usage{
JAGSmeta(
  formula,
  v,
  studyID,
  data,
  n_chains = 3,
  n_updates = 10000,
  n_iterations = 10000,
  prob_conditions = NULL
)
}
\arguments{
\item{formula}{Model formula that specifies the model}

\item{v}{Variance variable in data frame}

\item{studyID}{Column containing the study IDs that effects will be nested in}

\item{data}{Data frame containing the variables used in the model}

\item{n_chains}{Number of chains used to generate samples}

\item{n_updates}{Number of updates used to generate samples}

\item{n_iterations}{Number of iterations used to generate samples}

\item{prob_conditions}{List of probability statements to be estimated for specified parameters}
}
\value{
MCMC Graphical model
}
\description{
Function for estimating a hierarchical Bayesian Meta-regression using RJAGS
}
\examples{
\dontrun{
result <- JAGSmeta(
formula = yi ~ 1 + moderator_1 + moderator_2, # Model formula
v = "vi", # Column name for variance
studyID = "study_id", # Column name for study identifier
data = example_data, # The dataset
n_chains = 5, # Number of MCMC chains
n_updates = 100000, # Number of updates for burn-in
n_iterations = 100000 # Number of iterations for sampling
 prob_conditions = list(
"mu.alpha" = list(
  "greater" =c(0,.2, .5, .8)),
"sd.alpha" = list(
  "lesser" = .1,
  "between" = list(c(.1,.2), c(.2,.5)),
  "greater" = .5)
  )
)
}
}
