% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_betas.R
\name{rename_betas}
\alias{rename_betas}
\title{rename_betas}
\usage{
rename_betas(samples, predictor_vars)
}
\arguments{
\item{samples}{Sample distributions from an RJAGS model}

\item{predictor_vars}{Predictor variables specified in the model syntax of the JAGSmeta function}
}
\value{
Renamed coefficents for the model object
}
\description{
Renames parameters of an RJAGS to match the names of the original parameters specified in the model syntax
}
\examples{
rename_betas(samples, predictor_vars)
}
