% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAGS.summary.R
\name{JAGS.summary}
\alias{JAGS.summary}
\title{JAGS.summary}
\usage{
JAGS.summary(mcmc_samples)
}
\arguments{
\item{mcmc_samples}{A RJAGS model Object}
}
\value{
A simplified model summary that excludes estimates at the study level
}
\description{
Creates a simplified model summary output
}
\examples{
\dontrun{

JAGS.summary(JAGSmeta_model)

}
}
